+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 ## [Unreleased]
 ### Added
 - Added refresh token functionality.
+- Added logout functionality.
 ### Changed
 - For changes in existing functionality.
 ### Deprecated